+++ Sources/TuistKit/Commands/Cache/CachePrintHashesCommand.swift
@@ -11,7 +11,8 @@ struct CachePrintHashesCommand: ParsableCommand {
 
     @Option(
         name: .shortAndLong,
-        help: "The path where the project will be generated."
+        help: "The path where the project will be generated.",
+        completion: .directory
     )
     var path: String?