+++ website/src/components/footer.jsx
@@ -59,17 +59,6 @@ export default () => {
               </Link>
             </div>
             <div className="px-5 py-2">
-              <Link
-                to="/faq"
-                alt="Opens the frequently asked questions section"
-                target="__blank"
-                className="text-base leading-6"
-                sx={{ color: 'gray', ':hover': { color: 'primary' } }}
-              >
-                FAQ
-              </Link>
-            </div>
-            <div className="px-5 py-2">
               <a
                 href={releasesUrl}
                 alt="Opens the releases page."